+++ lib/src/fragment.dart
@@ -1,26 +1,25 @@
 import 'package:flutter/widgets.dart';
 import 'package:fragment/src/utils.dart';
 
-/// A widget to cache [builder] with [deps]
-/// The builder will be called only when [deps] is different (not shallowly equal)
-/// from the previous [deps].
+/// A widget to cache [builder] with [keys]
+/// The builder will be called only when [keys] is different (not shallowly equal)
+/// from the previous [keys].
 class Fragment<T extends Widget> extends StatefulWidget {
   final T Function(BuildContext context, T prev) builder;
-  final Iterable deps;
+  final Iterable keys;
 
-  const Fragment(this.builder, {this.deps = const [], Key key})
-      : super(key: key);
+  const Fragment(this.builder, {this.keys = const []});
 
   @override
   operator ==(Object other) {
     if (other is Fragment<T>) {
-      return shallowEquals(deps, other.deps);
+      return shallowEquals(keys, other.keys);
     }
     return false;
   }
 
   @override
-  int get hashCode => deps.hashCode;
+  int get hashCode => keys.hashCode;
 
   @override
   _FragmentState createState() => _FragmentState();